+++ config/jinja2.py
@@ -21,7 +21,7 @@ from wies.core.editables import (
 )
 from wies.core.inline_edit.jinja import inline_edit
 from wies.core.permission_engine import Verb, has_permission
-from wies.core.permissions import can_view_staff_page
+from wies.core.permissions import is_staff_member
 from wies.core.services.organizations import get_org_breadcrumb
 from wies.core.services.version import get_app_version
 
@@ -150,7 +150,7 @@ def environment(**options):
             "get_toggle_sort_url": get_toggle_sort_url,
             "get_sort_state": get_sort_state,
             "get_messages": get_messages,
-            "can_view_staff_page": can_view_staff_page,
+            "is_staff_member": is_staff_member,
             "DEBUG": settings.DEBUG,
             "APP_VERSION": get_app_version(),
             "inline_edit": inline_edit,